Update docs
I find it confusing to set a key name to "models" in the example config dictionary, when an entry name in it is required to be the same. Also I changed database name in example DB_URL string to "test" instead of "events", which I find more consistent with the example above it (dict format).