Validate JSON after decode
Closes #621

Custom JSON validations are currently run before the value stored in the database is converted to the Python type. This creates an asymmetry in the validation function as it runs against the Python type on create/update, but against the raw JSON string on fetch.

This changes the order of operations so the JSON is decoded before validation is run, along with a test for each of the three supported input types for the field.