Remove obsolete loop._selector
The loop._selector was introduced in commit 40f4945 with the ability to
provide loop to the initializer. At that point, the loop was
BaseSelectorEventLoop, which has a private _selector property.

In 743795d the BaseSelectorEventLoop was replaced by AbstractEventLoop,
but the _selector artefact remains.

The problem can be easily identified, when you use uvloop, which
implements AbstractEventLoop and is not a SelectorEventLoop.

Resolves: #532
See also: #533