Add value type hints to built-in fields
As the `Field` class isn't used as a descriptor (values are applied
directly to instances of models), type checkers see attributes typed as
the field instances themselves, rather than the types they represent.

This change makes `Field` a generic class, and implements stub `__get__`
and `__set__` methods such that the field type will match the value it's
replaced with, when accessed via an instance.